Moved routers to package-style subdirs
